Prevent players from using the /fix command from essentials on custom items. It even takes over completely for hand-only repairs

+++ plug-in/src/main/java/nl/knokko/customitems/plugin/set/item/CustomTool.java
@@ -280,7 +280,7 @@ public class CustomTool extends CustomItem {
 		}
 	}
 	
-	public IncreaseDurabilityResult increaseDurability(ItemStack stack, int amount) {
+	public IncreaseDurabilityResult increaseDurability(ItemStack stack, long amount) {
 		if (isUnbreakable() || !stack.hasItemMeta()) {
 			return new IncreaseDurabilityResult(stack, 0);
 		}
